requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
Django==5.2.11
djangorestframework==3.16.1
mysqlclient==2.2.7
redis==7.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
